fixes middleware roles on routes

+++ routes/web.php
@@ -27,32 +27,32 @@ Route::middleware('guest')->group(function () {
 });
 
 
-Route::middleware('auth.role:administrator')->group(function () {
+//Route::middleware('auth.role:administrator')->group(function () {
 
-        //classrooms
-        Route::get('classrooms/create', [App\Http\Controllers\ClassroomController::class, 'create'])->name('classrooms.create');
-        Route::post('classrooms', [App\Http\Controllers\ClassroomController::class, 'store'])->name('classrooms.store');
-        Route::put('classrooms/{classroom}', [App\Http\Controllers\ClassroomController::class, 'update'])->name('classrooms.update');
-        Route::get('classrooms/{classroom}/edit', [App\Http\Controllers\ClassroomController::class, 'edit'])->name('classrooms.edit');
-        Route::delete('classrooms/{classroom}', [App\Http\Controllers\ClassroomController::class, 'destroy'])->name('classrooms.destroy');
+    //classrooms
+    Route::get('classrooms/create', [App\Http\Controllers\ClassroomController::class, 'create'])->name('classrooms.create')->middleware('auth.role:educator, administrator');
+    Route::post('classrooms', [App\Http\Controllers\ClassroomController::class, 'store'])->name('classrooms.store')->middleware('auth.role:educator, administrator');
+    Route::put('classrooms/{classroom}', [App\Http\Controllers\ClassroomController::class, 'update'])->name('classrooms.update')->middleware('auth.role:educator, administrator');
+    Route::get('classrooms/{classroom}/edit', [App\Http\Controllers\ClassroomController::class, 'edit'])->name('classrooms.edit')->middleware('auth.role:educator, administrator');
+    Route::delete('classrooms/{classroom}', [App\Http\Controllers\ClassroomController::class, 'destroy'])->name('classrooms.destroy')->middleware('auth.role:educator, administrator');
 
-        //courses
-        Route::get('courses/create', [App\Http\Controllers\CourseController::class, 'create'])->name('courses.create');
-        Route::post('courses', [App\Http\Controllers\CourseController::class, 'store'])->name('courses.store');
-        Route::put('courses/{course}', [App\Http\Controllers\CourseController::class, 'update'])->name('courses.update');
-        Route::get('courses/{course}/edit', [App\Http\Controllers\CourseController::class, 'edit'])->name('courses.edit');
-        Route::delete('courses/{course}', [App\Http\Controllers\CourseController::class, 'destroy'])->name('courses.destroy');
+    //courses
+    Route::get('courses/create', [App\Http\Controllers\CourseController::class, 'create'])->name('courses.create')->middleware('auth.role:educator, administrator');
+    Route::post('courses', [App\Http\Controllers\CourseController::class, 'store'])->name('courses.store')->middleware('auth.role:educator, administrator');
+    Route::put('courses/{course}', [App\Http\Controllers\CourseController::class, 'update'])->name('courses.update')->middleware('auth.role:educator, administrator');
+    Route::get('courses/{course}/edit', [App\Http\Controllers\CourseController::class, 'edit'])->name('courses.edit')->middleware('auth.role:educator, administrator');
+    Route::delete('courses/{course}', [App\Http\Controllers\CourseController::class, 'destroy'])->name('courses.destroy')->middleware('auth.role:educator, administrator');
 
-        //attach courses to classroom
-        Route::post('classroom/courses/{id}/create', [App\Http\Controllers\ClassroomCoursesController::class, 'create']);
-        Route::delete('classroom/courses/{id}', [App\Http\Controllers\ClassroomCoursesController::class, 'destroy']);
+    //attach courses to classroom
+    Route::post('classroom/courses/{id}/create', [App\Http\Controllers\ClassroomCoursesController::class, 'create'])->middleware('auth.role:educator, administrator');
+    Route::delete('classroom/courses/{id}', [App\Http\Controllers\ClassroomCoursesController::class, 'destroy'])->middleware('auth.role:educator, administrator');
 
-        Route::get('upload-file/create', [App\Http\Controllers\TheoryController::class, 'createForm'])->name('create.pdf.form');
-        Route::post('upload-file', [App\Http\Controllers\TheoryController::class, 'fileUpload'])->name('upload.pdf');
-});
+    Route::get('upload-file/create', [App\Http\Controllers\TheoryController::class, 'createForm'])->name('create.pdf.form')->middleware('auth.role:educator, administrator');
+    Route::post('upload-file', [App\Http\Controllers\TheoryController::class, 'fileUpload'])->name('upload.pdf')->middleware('auth.role:educator, administrator');
+//});
 
 
-Route::middleware('auth.role:educator')->group(function () {
+/*Route::middleware('auth.role:educator')->group(function () {
 
     //classrooms
     Route::get('classrooms/create', [App\Http\Controllers\ClassroomController::class, 'create'])->name('classrooms.create');
@@ -63,9 +63,7 @@ Route::middleware('auth.role:educator')->group(function () {
     Route::delete('classrooms/{classroom}', [App\Http\Controllers\ClassroomController::class, 'destroy'])->name('classrooms.destroy');
 
     //courses
-    //Route::get('courses', [App\Http\Controllers\CourseController::class, 'index'])->name('courses.index');
     Route::get('courses/create', [App\Http\Controllers\CourseController::class, 'create'])->name('courses.create');
-    //Route::get('courses/{course}', [App\Http\Controllers\CourseController::class, 'show'])->name('courses.show');
     Route::post('courses', [App\Http\Controllers\CourseController::class, 'store'])->name('courses.store');
     Route::put('courses/{course}', [App\Http\Controllers\CourseController::class, 'update'])->name('courses.update');
     Route::get('courses/{course}/edit', [App\Http\Controllers\CourseController::class, 'edit'])->name('courses.edit');
@@ -78,7 +76,7 @@ Route::middleware('auth.role:educator')->group(function () {
     Route::get('upload-file/create', [App\Http\Controllers\TheoryController::class, 'createForm'])->name('create.pdf.form');
     Route::post('upload-file', [App\Http\Controllers\TheoryController::class, 'fileUpload'])->name('upload.pdf');
 
-});
+});*/
 
 Route::middleware('auth')->group(function () {
     Route::get('logout', [App\Http\Controllers\Auth\LoginController::class, 'logout'])->name('auth.logout');
@@ -120,6 +118,9 @@ Route::middleware('auth')->group(function () {
     //attach courses to classroom
     Route::post('classroom/courses/{id}/create', [App\Http\Controllers\ClassroomCoursesController::class, 'create']);
     Route::delete('classroom/courses/{id}', [App\Http\Controllers\ClassroomCoursesController::class, 'destroy']);
+
+    Route::get('upload-file/create', [App\Http\Controllers\TheoryController::class, 'createForm'])->name('create.pdf.form');
+    Route::post('upload-file', [App\Http\Controllers\TheoryController::class, 'fileUpload'])->name('upload.pdf');
 
     //user profile
     Route::get('user/profile', [App\Http\Controllers\ProfileController::class, 'edit'])->name('profile.edit');
